agent.py
"""
Coding agent implementation for Backspace
"""

import os
import tempfile
import shutil
from typing import AsyncGenerator, Dict, Any
import git
from github import Github
import asyncio
import json

class CodingAgent:
    """Handles code analysis and modification"""
    
    def __init__(self, github_token: str):
        self.github_token = github_token
        self.github = Github(github_token)
    
    async def process_repository(
        self, 
        repo_url: str, 
        prompt: str
    ) -> AsyncGenerator[Dict[str, Any], None]:
        """
        Process a repository and yield status updates
        """
        
        # Extract owner and repo name
        parts = repo_url.replace("https://github.com/", "").split("/")
        owner, repo_name = parts[0], parts[1]
        
        # Create temporary directory for cloning
        with tempfile.TemporaryDirectory() as temp_dir:
            try:
                # Clone repository
                yield {"type": "status", "stage": "clone", "message": f"Cloning {repo_url}..."}
                repo = git.Repo.clone_from(repo_url, temp_dir)
                yield {"type": "status", "stage": "clone", "message": "Repository cloned successfully"}
                
                # Analyze codebase
                yield {"type": "status", "stage": "analyze", "message": "Analyzing codebase structure..."}
                
                # List all Python files as an example
                py_files = []
                for root, dirs, files in os.walk(temp_dir):
                    # Skip .git directory
                    if '.git' in root:
                        continue
                    for file in files:
                        if file.endswith('.py'):
                            py_files.append(os.path.join(root, file))
                
                yield {
                    "type": "status", 
                    "stage": "analyze", 
                    "message": f"Found {len(py_files)} Python files"
                }
                
                # AI analysis and code modification workflow not yet implemented
                yield {
                    "type": "error", 
                    "message": "Code modification workflow not yet implemented. Current implementation supports repository cloning and analysis only."
                }
                
            except Exception as e:
                yield {
                    "type": "error",
                    "message": f"Error processing repository: {str(e)}"
                }
    
    async def _create_pull_request(self, repo_url: str, branch_name: str, prompt: str) -> str:
        """Create a pull request for the changes"""
        # Extract owner/repo from URL
        parts = repo_url.replace("https://github.com/", "").split("/")
        owner, repo_name = parts[0], parts[1]
        
        try:
            repo = self.github.get_repo(f"{owner}/{repo_name}")
            
            # Create pull request
            pr = repo.create_pull(
                title=f"Automated changes: {prompt}",
                body=f"This pull request implements the following changes:\n\n{prompt}\n\nGenerated automatically by the coding agent.",
                head=branch_name,
                base="main"
            )
            
            return pr.html_url
            
        except Exception as e:
            # Return error URL for now - in production this should handle gracefully
            return f"{repo_url}/pull/error"
    
    async def _analyze_and_plan_changes(self, repo_path: str, prompt: str) -> dict:
        """Analyze codebase and plan changes using AI"""
        raise NotImplementedError("AI integration not yet implemented. This method should integrate with OpenAI/Claude API.")
    
    async def _apply_code_changes(self, repo_path: str, changes: dict) -> None:
        """Apply the planned code changes to files"""
        raise NotImplementedError("Code modification not yet implemented. This method should apply changes to files based on AI recommendations.")
    
    async def _create_git_branch_and_commit(self, repo, branch_name: str, prompt: str) -> None:
        """Create git branch and commit changes"""
        raise NotImplementedError("Git operations not yet implemented. This method should create branches and commit changes.")


async def run_agent(repo_url: str, prompt: str) -> AsyncGenerator[str, None]:
    """
    Run the coding agent and yield SSE-formatted events
    """
    
    # Get GitHub token from environment
    github_token = os.getenv("GITHUB_TOKEN", "")
    
    if not github_token:
        yield f"data: {json.dumps({'type': 'error', 'message': 'GitHub token not configured'})}\n\n"
        return
    
    agent = CodingAgent(github_token)
    
    async for event in agent.process_repository(repo_url, prompt):
        yield f"data: {json.dumps(event)}\n\n"