requirements.txt
modal
fastapi
uvicorn
pydantic
httpx
GitPython
PyGithub
python-dotenv
ddtrace